Template - Harness Day-Zero Pipeline and Repository (#71)
* Template - Harness Day-Zero Pipeline and Repository

The execution of this template will create a new pipeline in the chosen project.  This pipeline will leverage the Stage Templates `sta_STO_SAST_SCA_Primer` and `CI_GoldenStandard_Container_Template`.  In addition to the creation of the pipeline, the code will add InputSets and Triggers as well as to create a new Harness Code Repository.

ADD: Template::ci-sto-hcr-standard:: pipeline, input_sets, triggers, and Harness Code Repository